refactor(ph-ai): extract sandbox logic from conversation viewset into executor module
Move ~370 lines of sandbox-specific code (constants, helpers, orchestration)
from ee/api/conversation.py into ee/hogai/sandbox/executor.py, replacing
the inline block with a single handle_sandbox_message() delegation call.
